Allows to send invitation and certificate in courses management page of an user
