+++ plugin/cursus/Resources/modules/DocumentModel/Controller/DocumentModelSelectionModalCtrl.js
@@ -31,8 +31,8 @@ export default class DocumentModelSelectionModalCtrl {
 
   submit () {
     if (this.documentModel) {
-      const url = Routing.generate('api_post_cursus_document_send', {documentModel: this.documentModel['id']})
-      this.$http.post(url, {sourceId: this.datas['id']}).then(d => {
+      const url = Routing.generate('api_post_cursus_document_send', {documentModel: this.documentModel['id'], sourceId: this.datas['id']})
+      this.$http.post(url).then(d => {
         if (d['status'] === 200) {
           if (this.callback) {
             this.callback(d['data'])
