+++ plugin/cursus/Resources/views/CursusRegistration/userSessionsRegistrationManagement.html.twig
@@ -102,12 +102,6 @@
 
 {% block javascripts %}
     {{ parent() }}
-    <script type="text/javascript"
-            src="{{ url('bazinga_jstranslation_js', { 'domain': 'cursus' }) }}"
-    >
-    </script>
-    <script src='{{ asset('bundles/clarolinecursus/js/userSessionsRegistrationManagement.js') }}'
-            type="text/javascript"
-    >
-    </script>
+    <script type="text/javascript" src="{{ url('bazinga_jstranslation_js', { 'domain': 'cursus' }) }}"></script>
+    <script src="{{ hotAsset('dist/claroline-distribution-plugin-cursus-user-registration-management.js') }}"></script>
 {% endblock %}
