+++ plugin/cursus/assets.json
@@ -2,7 +2,11 @@
   "webpack": {
     "entry": {
       "registration": "Registration/module",
-      "management": "Management/management"
+      "management": "Management/management",
+      "user-management": "User/userManagement",
+      "user-registration-management": "User/userRegistrationManagement",
+      "group-management": "Group/groupManagement",
+      "group-registration-management": "Group/groupRegistrationManagement"
     }
   }
 }
